apps/worker/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2024.3.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
llvmlite==0.42.0
load-dotenv==0.1.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai==1.14.2
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
pika==1.3.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
semantic-version==2.10.0
setuptools-rust==1.9.0
sniffio==1.3.1
sympy==1.12
tiktoken==0.6.0
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
